+++ reconstruction/reconstruction_window.h
@@ -38,7 +38,7 @@ private:
     unsigned char view_orientation = 2;
 private: //bad slices
     bool bad_slice_analzed = false;
-    std::vector<std::pair<int,int> > bad_slices;
+    std::vector<std::pair<size_t,size_t> > bad_slices;
 private:
     Ui::reconstruction_window *ui;
     std::shared_ptr<ImageModel> handle;
